requirements.txt
asgiref==3.5.2
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
tzdata==2022.1